libs/chain/pyproject.toml
[project]
name = "ai2i-chain"
version = "0.0.1"
requires-python = ">=3.12.8"
readme = "README.md"
dependencies = [
    "google-genai>=1.10.0",
    "httpx~=0.28.1",
    "langchain-core~=1.0",
    "langchain-openai~=1.0",
    "pydantic==2.10.4",
    "tenacity~=9.1.2",
    "langchain-classic>=1.0.0",
]

[tool.uv]
default-groups = ["types", "test", "tools"]

[dependency-groups]
types = [
    "microsoft-python-type-stubs",
]
test = [
    "brotli~=1.1.0",
    "deepmerge==2.0",
    "python-dotenv~=1.0.1",
    "pytest-asyncio~=0.25.0",
    "pytest-cov~=6.0.0",
    "pytest-env~=1.1.5",
    "pytest-mpl~=0.17.0",
    "pytest-snapshot~=0.9.0",
    "pytest-xdist~=3.6.1",
    "pytest~=8.3.4",
]
tools = [
    "ruff>=0.8.4",
    "flake8>=7.1.1",
    "tach>=0.21.0",
    "pyright~=1.1.398",
    "deptry~=0.23.0",
]

[tool.deptry]
extend_exclude = [
  ".*\\/test_.*\\.py",
  ".*conftest.py"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ai2i"]

[tool.pyright]
typeCheckingMode = "standard"
